% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: actual_uptake_tool}
\alias{: actual_uptake_tool}
\alias{actual_uptake_tool}
\title{: based on uptake of two of the three nutrients, yield is estimated being limited by the missing.}
\usage{
actual_uptake_tool(ds_supply)
}
\arguments{
\item{ds_supply}{: nutrient supply}
}
\description{
: part of QUEFTS
}
