% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: CIS_MarkdownText}
\alias{: CIS_MarkdownText}
\alias{CIS_MarkdownText}
\title{: process the CIS recom output to be used in the Markdown}
\usage{
CIS_MarkdownText(
  rr,
  fertilizers,
  userName,
  country,
  userPhoneNr,
  userField,
  area,
  areaUnits,
  PD,
  HD,
  email,
  lat,
  lon,
  rootUP,
  cassPD,
  cassUW,
  maxInv,
  userPhoneCC,
  sweetPotatoUP,
  sweetPotatoPD,
  tuberUP,
  sweetPotatoUW
)
}
\arguments{
\item{rr}{recommendation text as obtained from the get recommendation functions}

\item{fertilizers}{: data frame with type, nutreint content and cost of fertilizers}

\item{userName}{: user name}

\item{country}{: NG or TZ}

\item{userPhoneNr}{phone number without country code}

\item{userField}{: if user has a anme for his field}

\item{area}{field size}

\item{areaUnits}{ha or acre}

\item{PD}{planting date}

\item{HD}{harvest date}

\item{email}{user email}

\item{lat}{latitude}

\item{lon}{longitude}

\item{rootUP}{root price}

\item{cassPD}{cassava product root, flour, gari, ...}

\item{cassUW}{cassave unit wt}

\item{maxInv}{max investement}

\item{userPhoneCC}{user cuntry code for phone}

\item{sweetPotatoUP}{: sweetpotato unit price}

\item{sweetPotatoPD}{: swwetpotato product}

\item{tuberUP}{: tuber unit price}

\item{sweetPotatoUW}{: sweet potato unit wt}
}
\description{
: process the CIS recom output to be used in the Markdown
}
