% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: fertilizerAdviseTable}
\alias{: fertilizerAdviseTable}
\alias{fertilizerAdviseTable}
\title{:   function to generate the components of the markdown doc}
\usage{
fertilizerAdviseTable(FR, IC, country, areaUnits)
}
\arguments{
\item{FR}{TRUE or FALSE, based on if FR advice is requested}

\item{IC}{TRUE or FALSE, based on if SP advice is requested}

\item{country}{NG or TZ}

\item{areaUnits}{ha or acre}
}
\description{
:   function to generate the components of the markdown doc
}
