% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: fertilizerFunc}
\alias{: fertilizerFunc}
\alias{fertilizerFunc}
\title{: function to creat a data frame with fertilizers}
\usage{
fertilizerFunc(
  NPK201216available = TRUE,
  NPK201216CostperBag = NA,
  NPK201216BagWt = 50,
  ureaavailable = TRUE,
  ureaCostperBag = NA,
  ureaBagWt = 50,
  MOPavailable = TRUE,
  MOPCostperBag = NA,
  MOPBagWt = 50,
  DAPavailable = TRUE,
  DAPCostperBag = NA,
  DAPBagWt = 50,
  NPK201010available = TRUE,
  NPK201010CostperBag = NA,
  NPK201010BagWt = 50,
  NPK151515available = TRUE,
  NPK151515CostperBag = NA,
  NPK151515BagWt = 50,
  TSPavailable = TRUE,
  TSPCostperBag = NA,
  TSPBagWt = 50,
  NPK171717available = FALSE,
  NPK171717CostperBag = NA,
  NPK171717BagWt = 50,
  Nafakaavailable = TRUE,
  NafakaCostperBag = NA,
  NafakaBagWt = 50,
  CANavailable = TRUE,
  CANCostperBag = NA,
  CANBagWt = 50,
  SSPavailable = TRUE,
  SSPCostperBag = NA,
  SSPBagWt = 50,
  YaraMila_UNIKavailable = TRUE,
  YaraMila_UNIKCostperBag = NA,
  YaraMila_UNIKBagWt = 50,
  newFert1name = NA,
  newFert1N_cont = NA,
  newFert1P2O5 = NA,
  newFert1K2O = NA,
  newFert1CostperBag = NA,
  newFert1BagWt = NA,
  newFert2name = NA,
  newFert2N_cont = NA,
  newFert2P2O5 = NA,
  newFert2K2O = NA,
  newFert2CostperBag = NA,
  newFert2BagWt = NA,
  newFert3name = NA,
  newFert3N_cont = NA,
  newFert3P2O5 = NA,
  newFert3K2O = NA,
  newFert3CostperBag = NA,
  newFert3BagWt = NA,
  newFert4name = NA,
  newFert4N_cont = NA,
  newFert4P2O5 = NA,
  newFert4K2O = NA,
  newFert4CostperBag = NA,
  newFert4BagWt = NA,
  newFert5name = NA,
  newFert5N_cont = NA,
  newFert5P2O5 = NA,
  newFert5K2O = NA,
  newFert5CostperBag = NA,
  newFert5BagWt = NA,
  country
)
}
\arguments{
\item{NPK201216available}{: TRUE or FALSE}

\item{NPK201216CostperBag}{: in local currency price for the bag wt}

\item{NPK201216BagWt}{: c(25,50,75,100)}

\item{ureaavailable}{: TRUE or FALSE}

\item{ureaCostperBag}{: in local currency price for the bag wt}

\item{ureaBagWt}{: c(25,50,75,100)}

\item{MOPavailable}{: TRUE or FALSE}

\item{MOPCostperBag}{: in local currency price for the bag wt}

\item{MOPBagWt}{: c(25,50,75,100)}

\item{DAPavailable}{: TRUE or FALSE}

\item{DAPCostperBag}{: in local currency price for the bag wt}

\item{DAPBagWt}{: c(25,50,75,100)}

\item{NPK201010available}{: TRUE or FALSE}

\item{NPK201010CostperBag}{: in local currency price for the bag wt}

\item{NPK201010BagWt}{: c(25,50,75,100)}

\item{NPK151515available}{: TRUE or FALSE}

\item{NPK151515CostperBag}{: in local currency price for the bag wt}

\item{NPK151515BagWt}{: c(25,50,75,100)}

\item{TSPavailable}{: TRUE or FALSE}

\item{TSPCostperBag}{: in local currency price for the bag wt}

\item{TSPBagWt}{: c(25,50,75,100)}

\item{NPK171717available}{: TRUE or FALSE}

\item{NPK171717CostperBag}{: in local currency price for the bag wt}

\item{NPK171717BagWt}{: c(25,50,75,100)}

\item{Nafakaavailable}{: TRUE or FALSE}

\item{NafakaCostperBag}{: in local currency price for the bag wt}

\item{NafakaBagWt}{: c(25,50,75,100)}

\item{CANavailable}{: TRUE or FALSE}

\item{CANCostperBag}{: in local currency price for the bag wt}

\item{CANBagWt}{: c(25,50,75,100)}

\item{SSPavailable}{: TRUE or FALSE}

\item{SSPCostperBag}{: in local currency price for the bag wt}

\item{SSPBagWt}{: c(25,50,75,100)}

\item{YaraMila_UNIKavailable}{: TRUE or FALSE}

\item{YaraMila_UNIKCostperBag}{: in local currency price for the bag wt}

\item{YaraMila_UNIKBagWt}{: c(25,50,75,100)}

\item{newFert1name}{: any name used by user}

\item{newFert1N_cont}{: the N content (in proportion)}

\item{newFert1P2O5}{: P2O5 in the fertilizer from which the P  = 0.44*P2O5}

\item{newFert1K2O}{: K2O in the fertilizer and K contenet = 0.83 * K2O}

\item{newFert1CostperBag}{: cost for a bag of fertilzer bag defined below}

\item{newFert1BagWt}{; is fertilizer sold per 25,50,75,100,... kg}

\item{newFert2name}{: any name used by user}

\item{newFert2K2O}{: K2O in the fertilizer and K contenet = 0.83 * K2O}

\item{newFert2CostperBag}{: cost for a bag of fertilzer bag defined below}

\item{newFert3name}{: any name used by user}

\item{newFert3K2O}{: K2O in the fertilizer and K contenet = 0.83 * K2O}

\item{newFert3CostperBag}{: cost for a bag of fertilzer bag defined below}

\item{newFert3BagWt}{; is fertilizer sold per 25,50,75,100,... kg}

\item{newFert4name}{: any name used by user}

\item{newFert4K2O}{: K2O in the fertilizer and K contenet = 0.83 * K2O}

\item{newFert4CostperBag}{: cost for a bag of fertilzer bag defined below}

\item{newFert4BagWt}{; is fertilizer sold per 25,50,75,100,... kg}

\item{newFert5name}{: any name used by user}

\item{newFert5K2O}{: K2O in the fertilizer and K contenet = 0.83 * K2O}

\item{newFert5CostperBag}{: cost for a bag of fertilzer bag defined below}

\item{newFert5BagWt}{; is fertilizer sold per 25,50,75,100,... kg}

\item{country}{MG or TZ}

\item{newFert2N_cont:}{the N content (in proportion)}

\item{newFert2P2O5:}{P2O5 in the fertilizer from which the P  = 0.44*P2O5}

\item{newFert2BagWt:}{; is fertilizer sold per 25,50,75,100,... kg}

\item{newFert3N_cont:}{the N content (in proportion)}

\item{newFert3P2O5:}{P2O5 in the fertilizer from which the P  = 0.44*P2O5}

\item{newFert4N_cont:}{the N content (in proportion)}

\item{newFert4P2O5:}{P2O5 in the fertilizer from which the P  = 0.44*P2O5}

\item{newFert5N_cont:}{the N content (in proportion)}

\item{newFert5P2O5:}{P2O5 in the fertilizer from which the P  = 0.44*P2O5}
}
\value{
: data frame with (type, N_cont, P_cont, K_cont, price) The NPK is elemental concentration and price is per kg of fertilizer
}
\description{
: function to creat a data frame with fertilizers
}
