% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: final_yield_tools}
\alias{: final_yield_tools}
\alias{final_yield_tools}
\title{: gives final possible yield}
\usage{
final_yield_tools(Uptake_Yield)
}
\arguments{
\item{Uptake_Yield:}{nutrient uptake and yield estimates}
}
\description{
: gives final possible yield
}
