% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: Function to obtain recommendations on cassava-maize intercropping.}
\alias{: Function to obtain recommendations on cassava-maize intercropping.}
\alias{getICrecommendations}
\title{:   Function to obtain recommendations on cassava-maize intercropping.}
\usage{
getICrecommendations(
  areaHa = 1,
  CMP = 1:5,
  cobUP,
  fertilizers,
  riskAtt = c(0, 1, 2)
)
}
\arguments{
\item{areaHa}{farm size}

\item{CMP}{c(1,2,3,4,5),  Current maize performance, score on a scale of 1 (very yellow and stunted) .. 5 (tall and dark green), NA if IC != TRUE and FR != TRUE, or NA if the user does not know (NA = default)}

\item{cobUP}{maize cob price}

\item{fertilizers}{data frame with fertilizer type, NPK and price}

\item{riskAtt}{c(0, 1, 2): Risk attitude of the farmer, with 0 being very risk-averse (low income farmers who cannot afford to loose on investment), 1 = risk-neutral and 2 = risk-loving (higher income farmers willing to take their chances for higher net returns)}
}
\value{
:     list of 2 dataframes: (i) cost benefit analysis for most profitable system, and (ii) fertilizer rates to apply.
}
\description{
:   Function to obtain recommendations on cassava-maize intercropping.
}
