% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: Function to obtain recommendations on cassava-sweet potato intercropping.}
\alias{: Function to obtain recommendations on cassava-sweet potato intercropping.}
\alias{getCISrecommendations}
\title{:  Function to obtain recommendations on cassava-sweet potato intercropping.}
\usage{
getCISrecommendations(
  areaHa = 1,
  FCY = 11,
  tuberUP,
  rootUP,
  fertilizers,
  riskAtt = c(0, 1, 2)
)
}
\arguments{
\item{areaHa}{ha or acre}

\item{FCY}{: farmers current yield}

\item{tuberUP}{: tuber price}

\item{rootUP}{: root price}

\item{fertilizers}{: data frame with type, nutreint content and cost of fertilizers}

\item{riskAtt}{: rist attitude of user}
}
\value{
:     list of 2 dataframes: (i) cost benefit analysis for most profitable system, and (ii) fertilizer rates to apply.
}
\description{
:  Function to obtain recommendations on cassava-sweet potato intercropping.
}
