% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{:   Function to obtain tillage recommendations (step 4 of 6 steps).}
\alias{:   Function to obtain tillage recommendations (step 4 of 6 steps).}
\alias{getPPrecommendations}
\title{: Function to obtain tillage recommendations (step 4 of 6 steps). Function to obtain recommendations on ploughing and ridging. Returns a dataframe with all possible combinations of}
\usage{
getPPrecommendations(
  areaHa,
  costLMO,
  ploughing,
  ridging,
  method_ploughing,
  method_ridging,
  FCY,
  rootUP
)
}
\arguments{
\item{areaHa}{farm size}

\item{costLMO}{a data frame with operation, method of land preparatin, cost and area}

\item{ploughing}{Logical, c(NA, TRUE, FALSE), indicating if the user conducts a ploughing operation in current practice, NA if PP != TRUE}

\item{ridging}{Logical, c(NA, TRUE, FALSE), indicating if the user conducts a harrowing operation in current practice, NA if PP != TRUE}

\item{method_ploughing}{Categorical: c("manual", "tractor", "N/A"), method of ploughing currently applied by the farmer. Note: "N/A" = not applicable becasue nPlough=0}

\item{method_ridging}{Categorical: c("manual", "tractor", "N/A"), method of ridging currently applied by the farmer. Note: "N/A" = not applicable becasue ridges=FALSE}

\item{FCY}{Farmer-reported current yield, in tonnes FM per ha (optional, default value = 11)}

\item{rootUP}{root price}
}
\value{
:   dataframe with cost benefit for various combinations of ploughing and ridging.
}
\description{
: Function to obtain tillage recommendations (step 4 of 6 steps). Function to obtain recommendations on ploughing and ridging. Returns a dataframe with all possible combinations of
}
