% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getCISrecText}
\alias{: getCISrecText}
\alias{getCISrecText}
\title{: process the recom output as Markdown input}
\usage{
getCISrecText(ds)
}
\arguments{
\item{ds}{: get recommendation output}
}
\description{
: process the recom output as Markdown input
}
