% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getFRrecommendations}
\alias{: getFRrecommendations}
\alias{getFRrecommendations}
\title{: the main fertilizer recommendation function}
\usage{
getFRrecommendations(
  lat,
  lon,
  pd,
  pw,
  HD,
  had,
  maxInv,
  fertilizers,
  rootUP,
  areaHa,
  country,
  FCY
)
}
\arguments{
\item{lat}{: latitude}

\item{lon}{: longitude}

\item{pd}{planting date in the xth day of the year format}

\item{pw}{planting week}

\item{HD}{harvest date}

\item{had}{difference between HD and PD}

\item{maxInv}{max investment}

\item{fertilizers}{a data frame with type, NPK contnet and price}

\item{rootUP}{root price}

\item{areaHa}{is area of land in ha#'}

\item{country}{should be NG or TZ}

\item{FCY}{based on user input five values based on user input are passed, the app converts the value per ha so it is always per ha that comes}

\item{return}{a data frame with lat,lon, plDate,N, P, K, WLY, CurrentY, TargetY, TC, NR, harvestDate and rates of fertilizer (if any)}
}
\description{
: the main fertilizer recommendation function
}
