% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getFRrecText}
\alias{: getFRrecText}
\alias{getFRrecText}
\title{: makes a text of the advice to be used in the R markdown report}
\usage{
getFRrecText(ds, country, fertilizers, rootUP)
}
\arguments{
\item{ds}{out put of get FR recommendation function}

\item{country}{NG or TZ}

\item{fertilizers}{a data frame with type, NPK contnet and price}

\item{rootUP}{root price}
}
\description{
: makes a text of the advice to be used in the R markdown report
}
