% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getICrecText}
\alias{: getICrecText}
\alias{getICrecText}
\title{: makes a text of the advice to be used in the R markdown report}
\usage{
getICrecText(ds, maizePD)
}
\arguments{
\item{ds}{out put of get IC recommendation function}

\item{maizePD}{maize product grain or cob}
}
\description{
: makes a text of the advice to be used in the R markdown report
}
