% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getPPrecText}
\alias{: getPPrecText}
\alias{getPPrecText}
\title{: makes a text of the advice to be used in the R markdown report}
\usage{
getPPrecText(ds, country = c("NG", "TZ"))
}
\arguments{
\item{ds}{out put of get PP recommendation function}

\item{country}{NG or TZ}
}
\description{
: makes a text of the advice to be used in the R markdown report
}
