% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getRecommendation for all use cases}
\alias{: getRecommendation for all use cases}
\alias{PP_MarkdownText}
\title{: PP_MarkdownText}
\usage{
PP_MarkdownText(
  userName,
  country,
  userPhoneNr,
  userField,
  area,
  areaUnits,
  PD,
  HD,
  email,
  lat,
  lon,
  rootUP,
  cassPD,
  cassUW,
  maxInv,
  ploughing,
  ridging,
  method_ploughing,
  method_ridging,
  userPhoneCC
)
}
\arguments{
\item{userName}{: user name}

\item{country}{: NG or TZ}

\item{userPhoneNr}{phone number without country code}

\item{userField}{: if user has a anme for his field}

\item{area}{field size}

\item{areaUnits}{ha or acre}

\item{PD}{planting date}

\item{HD}{harvest date}

\item{email}{user email}

\item{lat}{latitude}

\item{lon}{longitude}

\item{rootUP}{root price}

\item{cassPD}{cassava product root, flour, gari, ...}

\item{cassUW}{cassava unit wt}

\item{maxInv}{max investement}

\item{ploughing}{: true or false}

\item{ridging}{; true or false}

\item{method_ploughing}{: tractor or mannual}

\item{method_ridging}{: rtactor or mannual}

\item{userPhoneCC}{user cuntry code for phone}
}
\description{
process the PP recom output to be used in the Markdown
}
