% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getRecommendation}
\alias{: getRecommendation}
\alias{getRecommendation}
\title{: Function with all inputs required by the DST app, and calling use case-specific wrapper functions to calculated recommendations on IC, FR, PP and SP.
            This function also calls on functions to send recommendation reports by SMS and/or email.Wrapper function to calculate recommendations across all use cases.}
\usage{
getRecommendation(
  country = c("NG", "TZ"),
  lat,
  lon,
  area,
  areaUnits = c("acre", "ha", "m2"),
  IC = c(TRUE, FALSE),
  intercrop = c("maize", "sweetpotato"),
  FR = c(TRUE, FALSE),
  PP = c(TRUE, FALSE),
  SPP = c(TRUE, FALSE),
  SPH = c(TRUE, FALSE),
  PD,
  HD,
  PD_window = c(0, 1, 2),
  HD_window = c(0, 1, 2),
  fallowType = c("bush", "broad_leaves", "grass", "none"),
  fallowHeight = c(NA, 100, 150, 200),
  fallowGreen = c(TRUE, FALSE),
  problemWeeds = c(TRUE, FALSE),
  tractor_plough = c(TRUE, FALSE),
  tractor_harrow = c(TRUE, FALSE),
  tractor_ridger = c(TRUE, FALSE),
  cost_LMO_areaBasis = c("areaUnit", "areaField"),
  cost_tractor_ploughing = NA,
  cost_tractor_harrowing = NA,
  cost_tractor_ridging = NA,
  cost_manual_ploughing = NA,
  cost_manual_harrowing = NA,
  cost_manual_ridging = NA,
  cost_weeding1 = NA,
  cost_weeding2 = NA,
  ploughing = c(TRUE, FALSE),
  harrowing = c(TRUE, FALSE),
  ridging = c(TRUE, FALSE),
  method_ploughing = c("manual", "tractor", "N/A"),
  method_harrowing = c("manual", "tractor", "N/A"),
  method_ridging = c("manual", "tractor", "N/A"),
  FCY,
  CMP,
  saleSF = c(TRUE, FALSE),
  nameSF = c(NA, "AlliedAtlanticDistilleries", "MatnaStarch", "PsaltryMarketers",
    "PsaltryOutgrowers", "Greentech", "ThaiFarm", "FJS"),
  cassPD = c("roots", "chips", "flour", "gari"),
  cassUW = c(1, 50, 100, 1000),
  cassUP,
  cassUP_m1,
  cassUP_m2,
  cassUP_p1,
  cassUP_p2,
  maizePD = c("fresh_cob", "grain"),
  maizeUW = c(1, 50, 100),
  maizeUP,
  sweetPotatoPD = c("tubers", "flour"),
  sweetPotatoUW = NA,
  sweetPotatoUP = NA,
  maxInv = NA,
  SMS = c(TRUE, FALSE),
  email = c(TRUE, FALSE),
  userPhoneCC,
  userPhoneNr,
  userName,
  userEmail,
  userField,
  riskAtt = c(0, 1, 2),
  NPK201216available = c(TRUE, FALSE),
  NPK201216CostperBag = NA,
  NPK201216BagWt = 50,
  ureaavailable = c(TRUE, FALSE),
  ureaCostperBag = NA,
  ureaBagWt = 50,
  MOPavailable = c(TRUE, FALSE),
  MOPCostperBag = NA,
  MOPBagWt = 50,
  DAPavailable = c(TRUE, FALSE),
  DAPCostperBag = NA,
  DAPBagWt = 50,
  NPK201010available = c(TRUE, FALSE),
  NPK201010CostperBag = NA,
  NPK201010BagWt = 50,
  NPK151515available = c(TRUE, FALSE),
  NPK151515CostperBag = NA,
  NPK151515BagWt = 50,
  TSPavailable = c(TRUE, FALSE),
  TSPCostperBag = NA,
  TSPBagWt = 50,
  NPK171717available = c(TRUE, FALSE),
  NPK171717CostperBag = NA,
  NPK171717BagWt = 50,
  Nafakaavailable = c(FALSE, FALSE),
  NafakaCostperBag = NA,
  NafakaBagWt = 50,
  CANavailable = c(TRUE, FALSE),
  CANCostperBag = NA,
  CANBagWt = 50,
  SSPavailable = c(TRUE, FALSE),
  SSPCostperBag = NA,
  SSPBagWt = 50,
  newFert1name = c(TRUE, FALSE),
  newFert1N_cont = NA,
  newFert1P2O5 = NA,
  newFert1K2O = NA,
  newFert1CostperBag = NA,
  newFert1BagWt = NA,
  newFert2name = c(TRUE, FALSE),
  newFert2N_cont = NA,
  newFert2P2O5 = NA,
  newFert2K2O = NA,
  newFert2CostperBag = NA,
  newFert2BagWt = NA,
  newFert3name = c(TRUE, FALSE),
  newFert3N_cont = NA,
  newFert3P2O5 = NA,
  newFert3K2O = NA,
  newFert3CostperBag = NA,
  newFert3BagWt = NA,
  newFert4name = c(TRUE, FALSE),
  newFert4N_cont = NA,
  newFert4P2O5 = NA,
  newFert4K2O = NA,
  newFert4CostperBag = NA,
  newFert4BagWt = NA,
  newFert5name = c(TRUE, FALSE),
  newFert5N_cont = NA,
  newFert5P2O5 = NA,
  newFert5K2O = NA,
  newFert5CostperBag = NA,
  newFert5BagWt = NA
)
}
\arguments{
\item{:}{newFert5BagWt=NA}
}
\value{
:     Vector of with recommendation texts to display.
}
\description{
: Function with all inputs required by the DST app, and calling use case-specific wrapper functions to calculated recommendations on IC, FR, PP and SP.
            This function also calls on functions to send recommendation reports by SMS and/or email.Wrapper function to calculate recommendations across all use cases.
}
