% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getSPrecommendations}
\alias{: getSPrecommendations}
\alias{getSPrecommendations}
\title{: get advice for SP}
\usage{
getSPrecommendations(
  areaHa,
  country,
  lat,
  lon,
  PD,
  HD,
  PD_window,
  HD_window,
  saleSF,
  nameSF,
  FCY,
  rootUP,
  rootUP_m1,
  rootUP_m2,
  rootUP_p1,
  rootUP_p2
)
}
\arguments{
\item{areaHa}{famr size}

\item{country}{NG or TZ}

\item{lat}{latitude}

\item{lon}{longitude}

\item{PD}{planting date}

\item{HD}{harvest date}

\item{PD_window}{planting window}

\item{HD_window}{harvest window}

\item{saleSF}{Logical, c(NA, TRUE, FALSE), indicating if the user is selling roots to a registered starch factory at factory-fixed prices}

\item{nameSF}{c(NA, "AlliedAtlanticDistilleries", "MatnaStarch", "PsaltryMarketers", "PsaltryOutgrowers",    "Greentech", "ThaiFarm", "FJS"), Name of starch factory where roots will be sold, NA if saleSF = FALSE}

\item{FCY}{Farmer-reported current yield, in tonnes FM per ha (optional, default value = 11)}

\item{rootUP}{root price}

\item{rootUP_m1}{root price a month before harvest}

\item{rootUP_m2}{root price two months before harvest}

\item{rootUP_p1}{root price a month after harvest}

\item{rootUP_p2}{root price two months after harvest}
}
\description{
: get advice for SP
}
