% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getSPrecText}
\alias{: getSPrecText}
\alias{getSPrecText}
\title{: makes a text of the advice to be used in the R markdown report}
\usage{
getSPrecText(ds, country)
}
\arguments{
\item{ds}{out put of get SP recommendation function}

\item{country}{NG or TZ}
}
\description{
: makes a text of the advice to be used in the R markdown report
}
