% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getsupply}
\alias{: getsupply}
\alias{getsupply}
\title{:}
\usage{
getsupply(dss)
}
\arguments{
\item{dss}{data frame with location, and QUEFTS crop parameters}
}
\description{
:
}
\author{
ACAI
}
