% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: getWMrecommendations}
\alias{: getWMrecommendations}
\alias{getWMrecommendations}
\title{: Function to obtain recommendations on land clearing (slashing and spraying) based on decision tree in the paper-based tool}
\usage{
getWMrecommendations(
  fallowType = c(NA, "bush", "broad_leaves", "grass", "none"),
  fallowHeight = c(NA, 100, 150, 200),
  fallowGreen = c(NA, TRUE, FALSE),
  problemWeeds = c(NA, TRUE, FALSE)
)
}
\arguments{
\item{fallowType}{: Categorical: c("bush", "broad_leaves", "grass", "none"), type of fallow prior to land clearing}

\item{fallowHeight}{: Categorical: c(NA, 100, 150, 200), height of the fallow prior to clearing}

\item{fallowGreen}{Logical: c(TRUE, FALSE), indicating if the fallow is lush, fresh and green (TRUE) or withered, dry or dead (FALSE)}

\item{problemWeeds}{Logical:  c(TRUE, FALSE), indicating the presence of any problem weeds that need to be controlled by herbicide (Tithonia, Imperata,...)}
}
\value{
: dataframe with recommendations on whether to slash and/or to spray.
}
\description{
: Function to obtain recommendations on land clearing (slashing and spraying) based on decision tree in the paper-based tool
}
