% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: IC_MarkdownText}
\alias{: IC_MarkdownText}
\alias{IC_MarkdownText}
\title{: process the FR recom output to be used in the Markdown}
\usage{
IC_MarkdownText(
  rr,
  fertilizers,
  userName,
  country,
  userPhoneNr,
  userField,
  area,
  areaUnits,
  PD,
  HD,
  email,
  lat,
  lon,
  rootUP,
  cassPD,
  maxInv,
  userPhoneCC,
  CMP,
  maizeUW,
  maizePD,
  cassUW,
  maizeUP
)
}
\arguments{
\item{rr}{recommendation text as obtained from the get recommendation functions}

\item{fertilizers}{: data frame with type, nutreint content and cost of fertilizers}

\item{userName}{: user name}

\item{country}{: NG or TZ}

\item{userPhoneNr}{phone number without country code}

\item{userField}{: if user has a anme for his field}

\item{area}{field size}

\item{areaUnits}{ha or acre}

\item{PD}{planting date}

\item{HD}{harvest date}

\item{email}{user email}

\item{lat}{latitude}

\item{lon}{longitude}

\item{rootUP}{root price}

\item{cassPD}{cassava product root, flour, gari, ...}

\item{maxInv}{max investement}

\item{userPhoneCC}{user cuntry code for phone}

\item{CMP}{: current maize height}

\item{maizeUW}{: mauze unit wt}

\item{maizePD}{: maize product}

\item{cassUW}{: cassava unit wt}

\item{maizeUP}{: maize unit price}
}
\description{
process the recom output as Markdown input
}
