% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: is a function to convert root DM yield into root fresh matter yield (RFY)}
\alias{: is a function to convert root DM yield into root fresh matter yield (RFY)}
\alias{getRFY}
\title{: Function to predict root FM yield based on date of harvest and country, using data from gravimetric starch measurements conducted across ACAI trials.}
\usage{
getRFY(HD, RDY, country)
}
\arguments{
\item{HD}{harvest date (Date format)}

\item{RDY}{root dry matter yield (user's units)}

\item{country}{c("NG", "TZ")}
}
\value{
:     RFY: root fresh yield in the same units as root DM yield input
}
\description{
: Function to predict root FM yield based on date of harvest and country, using data from gravimetric starch measurements conducted across ACAI trials.
}
