% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: is a Function to convert root FM yield into root dry matter yield (RDY): user define CY in FM in ton/ha, QUEFTS require Cy in DM kg/ha}
\alias{: is a Function to convert root FM yield into root dry matter yield (RDY): user define CY in FM in ton/ha, QUEFTS require Cy in DM kg/ha}
\alias{getRDY}
\title{: Function to predict root FM yield based on date of harvest and country, using data from gravimetric starch measurements conducted across ACAI trials.}
\usage{
getRDY(HD, RFY, country)
}
\arguments{
\item{HD}{harvest date (Date format)#'}

\item{RFY}{root dry matter yield (user's units)}

\item{country}{c("NG", "TZ")}
}
\value{
:     RDY: root dry yield in the same units as root FM yield input
}
\description{
: Function to predict root FM yield based on date of harvest and country, using data from gravimetric starch measurements conducted across ACAI trials.
}
