% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: max_min_yields_tools}
\alias{: max_min_yields_tools}
\alias{max_min_yields_tools}
\title{based on npk uptake gives max and min yield}
\usage{
max_min_yields_tools(dss)
}
\arguments{
\item{dss}{a data frame with nutrient uptake}
}
\description{
based on npk uptake gives max and min yield
}
