% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: NRabove18Cost}
\alias{: NRabove18Cost}
\alias{NRabove18Cost}
\title{see if profit is > (0.18 * total cost) + total cost}
\usage{
NRabove18Cost(ds)
}
\arguments{
\item{ds}{: data frame with fertilzer rate cost and net revenue}
}
\description{
see if profit is > (0.18 * total cost) + total cost
}
