% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: NUE}
\alias{: NUE}
\alias{NUE}
\title{: nutrient use effciency}
\usage{
NUE(
  HI,
  CmaxNroots = 6.6,
  CminNroots = 2.5,
  CmaxNtops = 17.9,
  CminNtops = 7.9,
  CmaxProots = 1.5,
  CminProots = 0.8,
  CmaxPtops = 2.8,
  CminPtops = 0.9,
  CmaxKroots = 11,
  CminKroots = 2.8,
  CmaxKtops = 18.8,
  CminKtops = 3.4
)
}
\arguments{
\item{HI}{: harvest index}

\item{CmaxNroots}{: QUEFTS parameter}

\item{CminNroots}{: QUEFTS parameter}

\item{CmaxNtops}{: QUEFTS parameter}

\item{CminNtops}{: QUEFTS parameter}

\item{CmaxProots}{: QUEFTS parameter}

\item{CminProots}{: QUEFTS parameter}

\item{CmaxPtops}{: QUEFTS parameter}

\item{CminPtops}{: QUEFTS parameter}

\item{CmaxKroots}{: QUEFTS parameter}

\item{CminKroots}{: QUEFTS parameter}

\item{CmaxKtops}{: QUEFTS parameter}

\item{CminKtops}{: QUEFTS parameter}
}
\description{
: nutrient use effciency
}
