% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: nutrient_uptake}
\alias{: nutrient_uptake}
\alias{nutrient_uptake}
\title{:  Nutrient uptake depends on the soil supply of the nutrient and the supply of other nutrients}
\usage{
nutrient_uptake(
  S1 = NA,
  S2 = NA,
  d1 = NA,
  a1 = NA,
  d2 = NA,
  a2 = NA,
  r1 = NA,
  r2 = NA
)
}
\arguments{
\item{S1}{: soil supply for nutrient 1}

\item{S2}{: soil supply for nutrient 2}

\item{d1}{: QUEFTS Parameter}

\item{a1}{: QUEFTS Parameter}

\item{d2}{: QUEFTS Parameter}

\item{a2}{: QUEFTS Parameter}

\item{r1}{: QUEFTS Parameter}

\item{r2}{: QUEFTS Parameter}
}
\description{
:  Nutrient uptake depends on the soil supply of the nutrient and the supply of other nutrients
}
