% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: optim_NR}
\alias{: optim_NR}
\alias{optim_NR}
\title{:   Optimize the UREA, TSP and MOP needed to maximize the NR. x1, x2, x3 = Urea, MOP and Nafaka kg/ha.}
\usage{
optim_NR(fertRate, rootUP, QID, CY, fertilizer, invest, HD, country)
}
\arguments{
\item{fertRate}{: NPK rate as defined by optimization}

\item{rootUP}{: cassava root price}

\item{QID}{: soil data}

\item{CY}{: current yield in DM in kg/ha}

\item{fertilizer}{: a data frame with type, NPK content and price}

\item{invest}{: investment capacity}

\item{HD}{; harvest date}

\item{country}{; NG or TZ}
}
\value{
: optimized fertilzer rates and yield
}
\description{
:   Optimize the UREA, TSP and MOP needed to maximize the NR. x1, x2, x3 = Urea, MOP and Nafaka kg/ha.
}
