% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: PPSP_MarkdownText}
\alias{: PPSP_MarkdownText}
\alias{PPSP_MarkdownText}
\title{: process the PP recom output to be used in the Markdown}
\usage{
PPSP_MarkdownText(
  rr,
  fname,
  userName = userName,
  country = country,
  userPhoneNr = userPhoneNr,
  userField = userField,
  area = area,
  areaUnits = areaUnits,
  PD = PD,
  HD = HD,
  email = email,
  lat = lat,
  lon = lon,
  rootUP = rootUP,
  cassPD = cassPD,
  cassUW = cassUW,
  maxInv = maxInv
)
}
\arguments{
\item{rr}{recommendation text as obtained from the get recommendation functions}

\item{fname}{: file name for output}

\item{userName}{: user name}

\item{country}{: NG or TZ}

\item{userPhoneNr}{phone number without country code}

\item{userField}{: if user has a anme for his field}

\item{area}{field size}

\item{areaUnits}{ha or acre}

\item{PD}{planting date}

\item{HD}{harvest date}

\item{email}{user email}

\item{lat}{latitude}

\item{lon}{longitude}

\item{rootUP}{root price}

\item{cassPD}{cassava product root, flour, gari, ...}

\item{cassUW}{casasva unit wt}

\item{maxInv}{max investement for fertilizers}
}
\description{
: process the PP recom output to be used in the Markdown
}
