% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: quefts_tools}
\alias{: quefts_tools}
\alias{quefts_tools}
\title{: Final yield based on the combinations of nutrient uptake and minimum + maximum yields.}
\usage{
quefts_tools(supply_wly)
}
\arguments{
\item{supply_wly:}{: QUEFTS Parameter}
}
\description{
: Final yield based on the combinations of nutrient uptake and minimum + maximum yields.
}
