% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: QUEFTS_WLY_CY}
\alias{: QUEFTS_WLY_CY}
\alias{QUEFTS_WLY_CY}
\title{The soil NPK as obtained from randdom forest model}
\usage{
QUEFTS_WLY_CY(SoilData = SoilData, country = country, wlyd = wlyd)
}
\arguments{
\item{SoilData}{: isric data}

\item{country}{: NG or TZ}

\item{wlyd}{a data frae wit Wwater limited yield, current yield and location#'}
}
\value{

}
\description{
The soil NPK as obtained from randdom forest model
}
\author{
ACAI
}
