% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: Rerun_25kgKa_try}
\alias{: Rerun_25kgKa_try}
\alias{Rerun_25kgKa_try}
\title{: after setting fertilizer recommendation <25 kg/ha Urea, MOP or Nafaka, target yield with the remaining recommended fertilizer is  re-estimated  and
 total cost, gross and net revenue are re calcuated.}
\usage{
Rerun_25kgKa_try(
  rootUP,
  rdd,
  fertilizer,
  QID,
  onlyFert,
  country,
  WLY = WLY,
  DCY = DCY,
  HD = HD,
  areaHa = areaHa
)
}
\arguments{
\item{rootUP}{cassava root price}

\item{rdd}{has fresh wt}

\item{fertilizer}{a data frame with type and cost of fertilizers}

\item{QID}{soil data}

\item{onlyFert}{recommended fertilizers}

\item{country}{NG or TZ}

\item{WLY}{water limited yield}

\item{DCY}{Current yield DM in kg/ha}

\item{HD}{harvest date}

\item{areaHa}{farm size in ha or acre}
}
\description{
: after setting fertilizer recommendation <25 kg/ha Urea, MOP or Nafaka, target yield with the remaining recommended fertilizer is  re-estimated  and
 total cost, gross and net revenue are re calcuated.
}
\author{
Meklit
}
