% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: Rfmodel_Wrapper}
\alias{: Rfmodel_Wrapper}
\alias{Rfmodel_Wrapper}
\title{: RF model works only if the factr levels are exactly identical to the data used to develp the model,}
\usage{
Rfmodel_Wrapper(FCY, country, lat, lon)
}
\arguments{
\item{FCY}{: farmers current yield}

\item{country}{: NG or TZ}

\item{lat}{: latitude}

\item{lon}{: lngitude
@export}
}
\description{
: RF model works only if the factr levels are exactly identical to the data used to develp the model,
}
