% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: run_Optim_NG2}
\alias{: run_Optim_NG2}
\alias{run_Optim_NG2}
\title{: get optimized fertilizer rate, target yield for the recommended rate and net revenue given cost and investment}
\usage{
run_Optim_NG2(
  rootUP,
  QID,
  fertilizer,
  invest,
  plDate,
  WLYData,
  lat,
  lon,
  areaHa,
  HD,
  WLY,
  DCY,
  country
)
}
\arguments{
\item{rootUP}{root price}

\item{QID}{soil data}

\item{fertilizer}{fertilizer data frame}

\item{invest}{capital to invest on fertlizer}

\item{plDate}{planting date}

\item{WLYData}{waterlimited yield}

\item{lat}{latitude}

\item{lon}{longitude}

\item{areaHa}{farm size}

\item{HD}{harvest date}

\item{WLY}{water limit yield}

\item{DCY}{dry current yield}

\item{country}{NG and TZ}
}
\value{
: a data frame with optimized fertilizer rates, TY and prices
}
\description{
: get optimized fertilizer rate, target yield for the recommended rate and net revenue given cost and investment
}
