% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: sendEmailReport}
\alias{: sendEmailReport}
\alias{sendEmailReport}
\title{: function to send mail}
\usage{
sendEmailReport(
  userEmail,
  FR,
  IC,
  PP,
  SP,
  FRrecom,
  ICrecom,
  country,
  PPrecom,
  SPrecom
)
}
\arguments{
\item{FR}{: TRUE or FALSE, based on if FR advice is requested used to generate the markdown doc}

\item{IC}{: TRUE or FALSE, based on if IC advice is requested used to generate the markdown doc}

\item{PP}{: TRUE or FALSE, based on if PP advice is requested used to generate the markdown doc}

\item{SP}{: TRUE or FALSE, based on if SP advice is requested used to generate the markdown doc}

\item{FRrecom}{: TRUE or FALSE, based on if FR advice is requested used to bundel together the pdf generated}

\item{ICrecom}{: TRUE or FALSE, based on if FR advice is requested used to bundel together the pdf generated}

\item{country}{: TRUE or FALSE, based on if FR advice is requested used to bundel together the pdf generated}

\item{PPrecom}{: TRUE or FALSE, based on if FR advice is requested used to bundel together the pdf generated}

\item{SPrecom}{: TRUE or FALSE, based on if FR advice is requested used to bundel together the pdf generated}

\item{userEmail:}{user email}
}
\description{
: function to send mail
}
