% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: SP_MarkdownText}
\alias{: SP_MarkdownText}
\alias{SP_MarkdownText}
\title{: process the SP recom output to be used in the Markdown}
\usage{
SP_MarkdownText(
  userName,
  country,
  userPhoneNr,
  userField,
  area,
  areaUnits,
  PD,
  HD,
  email,
  lat,
  lon,
  saleSF,
  nameSF,
  maxInv,
  ploughing,
  ridging,
  method_ploughing,
  method_ridging,
  userPhoneCC,
  CMP,
  riskAtt,
  PD_window,
  HD_window,
  cassPD,
  cassUW,
  cassUP,
  cassUP_m1,
  cassUP_m2,
  cassUP_p1,
  cassUP_p2
)
}
\arguments{
\item{userName}{: user name}

\item{country}{: NG or TZ}

\item{userPhoneNr}{phone number without country code}

\item{userField}{: if user has a anme for his field}

\item{area}{field size}

\item{areaUnits}{ha or acre}

\item{PD}{planting date}

\item{HD}{harvest date}

\item{email}{user email}

\item{lat}{latitude}

\item{lon}{longitude}

\item{saleSF}{: saling cassava for factory true or false}

\item{nameSF}{; name of cassava processing factroy}

\item{maxInv}{: max investment for fertilizers}

\item{ploughing}{: true or false}

\item{ridging}{: true or false}

\item{method_ploughing}{: tractor or mannual}

\item{method_ridging}{: tractor or mannual}

\item{userPhoneCC}{: user phone number country code}

\item{CMP}{: current maize height}

\item{riskAtt}{: reis attitude of user}

\item{PD_window}{:  planting window}

\item{HD_window}{: harvest window}

\item{cassPD}{: cassava product}

\item{cassUW}{: cassava unit wt}

\item{cassUP}{: cassava price}

\item{cassUP_m1}{: cassava price one month from the harvest data in the future}

\item{cassUP_m2}{: cassava price two months from the harvest data in the future}

\item{cassUP_p1}{: cassava price one month from the harvest data in the past}

\item{cassUP_p2}{: cassava price two months from the harvest data in the past}
}
\description{
: process the SP recom output to be used in the Markdown
}
