% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: water_dependent_nutrient_uptake}
\alias{: water_dependent_nutrient_uptake}
\alias{water_dependent_nutrient_uptake}
\title{: get WLY limited nutrient uptake}
\usage{
water_dependent_nutrient_uptake(S1 = NA, WLY = NA, d1 = NA, a1 = NA, r1 = NA)
}
\arguments{
\item{S1}{: soil supply for nutrient 1}

\item{WLY}{: water limited yield}

\item{d1}{: QUEFTS Parameter}

\item{a1}{: QUEFTS Parameter}

\item{r1}{: QUEFTS Parameter}
}
\description{
: get WLY limited nutrient uptake
}
