% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkilimoFunctions_dev.R
\name{: yield_nutrients_combined}
\alias{: yield_nutrients_combined}
\alias{yield_nutrients_combined}
\title{: Yield calculated based on the combined uptake of 2 nutrients, while taking into account the availability of the third nutrient.}
\usage{
yield_nutrients_combined(
  U1 = NA,
  d1 = NA,
  a1 = NA,
  Y2A = NA,
  Y2D = NA,
  Y3D = NA,
  r1 = NA
)
}
\arguments{
\item{U1}{: QUEFTS Parameter}

\item{d1}{: QUEFTS Parameter}

\item{a1}{: QUEFTS Parameter}

\item{Y2A}{: QUEFTS Parameter}

\item{Y2D}{: QUEFTS Parameter}

\item{Y3D}{: QUEFTS Parameter}

\item{r1}{: QUEFTS Parameter}
}
\description{
: Yield calculated based on the combined uptake of 2 nutrients, while taking into account the availability of the third nutrient.
}
